Deleted to many files
